Fix Metrics/ClassLength
